+++ src/main/java/com/training/courier/exception/DeliveryNotFoundException.java
@@ -1,0 +1,7 @@
+package com.training.courier.exception;
+
+public class CourierDeliveryNotFoundException extends RuntimeException{
+    public CourierDeliveryNotFoundException(Long id) {
+        super("Courier delivery with " + id + " not found");
+    }
+}
