+++ src/main/java/com/training/courier/exception/SalaryNotFoundException.java
@@ -1,0 +1,7 @@
+package com.training.courier.exception;
+
+public class NotFoundException extends RuntimeException{
+    public NotFoundException(Long id) {
+        super("Entity with " + id + " not found");
+    }
+}
