+++ src/main/java/com/training/courier/service/task/DeliveriesSyncTask.java
@@ -1,0 +1,2 @@
+package com.training.courier.service.task;public interface DeliveriesSyncTask {
+}
